requirements.txt
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Django==4.2
djangorestframework==3.14.0
executing==1.2.0
ipdb==0.13.13
ipython==8.12.0
jedi==0.18.2
kombu==5.2.4
matplotlib-inline==0.1.6
mysqlclient==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pytz==2023.3
redis==4.5.4
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.6
git+https://github.com/jeagercoder/service_lib.git@v0.0.3
